+++ src/server/graphql/blockchain/blockchain.ts
@@ -181,6 +181,7 @@ export const resolvers: Resolvers<QRequestContext> = {
                     context,
                     info,
                     traceSpan,
+                    args.archive,
                 )
             })
         },
@@ -193,6 +194,7 @@ export const resolvers: Resolvers<QRequestContext> = {
                         context,
                         info,
                         traceSpan,
+                        args.archive,
                     )
                 },
             )
@@ -204,6 +206,7 @@ export const resolvers: Resolvers<QRequestContext> = {
                     context,
                     info,
                     traceSpan,
+                    args.archive,
                 )
             })
         },
@@ -214,6 +217,7 @@ export const resolvers: Resolvers<QRequestContext> = {
                     context,
                     info,
                     traceSpan,
+                    args.archive,
                 )
             })
         },