+++ src/server/graphql/blockchain/fetchers/messages.ts
@@ -22,16 +22,28 @@ import {
     BlockchainMessageTypeFilterEnum,
 } from "../resolvers-types-generated"
 import { resolveAddress } from "../../../address"
+import {
+    upgradeSelectionForBocParsing,
+    messageArchiveFields,
+    parseMessageBocsIfRequired,
+} from "../boc-parsers"
+import { useMessagesArchive } from "../../../data/data-provider"
 
 export async function resolve_message(
     hash: String,
     context: QRequestContext,
     info: GraphQLResolveInfo,
     traceSpan: QTraceSpan,
+    archive: boolean | undefined | null,
 ) {
     const maxJoinDepth = 1
+    const useArchive = useMessagesArchive(archive, context)
+    const { selectionSet, requireBocParsing } = upgradeSelectionForBocParsing(
+        useArchive,
+        info.fieldNodes[0].selectionSet,
+        messageArchiveFields,
+    )
 
-    const selectionSet = info.fieldNodes[0].selectionSet
     const returnExpression = config.messages.buildReturnExpression(
         selectionSet,
         context,
@@ -48,16 +60,21 @@ export async function resolve_message(
         "FOR doc IN messages " +
         `FILTER doc._key == @${params.add(hash)} ` +
         `RETURN ${returnExpression}`
-    const queryResult = (await context.services.data.query(
-        required(context.services.data.messages.provider),
-        {
-            text: query,
-            vars: params.values,
-            orderBy: [],
-            request: context,
-            traceSpan,
-        },
-    )) as BlockchainMessage[]
+    const queryResult = await parseMessageBocsIfRequired(
+        requireBocParsing,
+        context,
+        (await context.services.data.query(
+            required(context.services.data.messages.provider),
+            {
+                text: query,
+                vars: params.values,
+                orderBy: [],
+                request: context,
+                traceSpan,
+                archive: useArchive,
+            },
+        )) as BlockchainMessage[],
+    )
 
     await config.messages.fetchJoins(
         queryResult,
@@ -120,8 +137,13 @@ export async function resolve_account_messages(
             }
         }
     }
+    const useArchive = useMessagesArchive(args.archive, context)
+    const { selectionSet, requireBocParsing } = upgradeSelectionForBocParsing(
+        useArchive,
+        getNodeSelectionSetForConnection(info),
+        messageArchiveFields,
+    )
 
-    const selectionSet = getNodeSelectionSetForConnection(info)
     const returnExpressionBuilder = (sortField: string) =>
         config.messages.buildReturnExpression(
             selectionSet,
@@ -252,22 +274,26 @@ export async function resolve_account_messages(
               } ` +
               `LIMIT ${limit} ` +
               "RETURN d"
-    const queryResult = (await context.services.data.query(
-        required(context.services.data.transactions.provider),
-        {
-            text: query,
-            vars: params.values,
-            orderBy: [
-                {
-                    path: "account_chain_order",
-                    direction: "ASC",
-                },
-            ],
-            distinctBy: "account_chain_order",
-            request: context,
-            traceSpan,
-            archive: args.archive ?? undefined,
-        },
+    const queryResult = (await parseMessageBocsIfRequired(
+        requireBocParsing,
+        context,
+        (await context.services.data.query(
+            required(context.services.data.transactions.provider),
+            {
+                text: query,
+                vars: params.values,
+                orderBy: [
+                    {
+                        path: "account_chain_order",
+                        direction: "ASC",
+                    },
+                ],
+                distinctBy: "account_chain_order",
+                request: context,
+                traceSpan,
+                archive: useArchive,
+            },
+        )) as BlockchainMessage[],
     )) as (BlockchainMessage & { account_chain_order?: string })[]
 
     return (await processPaginatedQueryResult(