+++ src/server/graphql/blockchain/fetchers/transactions.ts
@@ -21,16 +21,29 @@ import {
     BlockchainTransaction,
     BlockchainTransactionsConnection,
 } from "../resolvers-types-generated"
+import {
+    parseTransactionBocsIfRequired,
+    transactionArchiveFields,
+    upgradeSelectionForBocParsing,
+} from "../boc-parsers"
+import { useTransactionsArchive } from "../../../data/data-provider"
 
 export async function resolve_transaction(
     hash: String,
     context: QRequestContext,
     info: GraphQLResolveInfo,
     traceSpan: QTraceSpan,
+    archive: boolean | undefined | null,
 ) {
     const maxJoinDepth = 2
 
-    const selectionSet = info.fieldNodes[0].selectionSet
+    const useArchive = useTransactionsArchive(archive, context)
+    const { selectionSet, requireBocParsing } = upgradeSelectionForBocParsing(
+        useArchive,
+        info.fieldNodes[0].selectionSet,
+        transactionArchiveFields,
+    )
+
     const returnExpression = config.transactions.buildReturnExpression(
         selectionSet,
         context,
@@ -47,16 +60,21 @@ export async function resolve_transaction(
         "FOR doc IN transactions " +
         `FILTER doc._key == @${params.add(hash)} ` +
         `RETURN ${returnExpression}`
-    const queryResult = (await context.services.data.query(
-        required(context.services.data.transactions.provider),
-        {
-            text: query,
-            vars: params.values,
-            orderBy: [],
-            request: context,
-            traceSpan,
-        },
-    )) as BlockchainTransaction[]
+    const queryResult = await parseTransactionBocsIfRequired(
+        requireBocParsing,
+        context,
+        (await context.services.data.query(
+            required(context.services.data.transactions.provider),
+            {
+                text: query,
+                vars: params.values,
+                orderBy: [],
+                request: context,
+                traceSpan,
+                archive: useArchive,
+            },
+        )) as BlockchainTransaction[],
+    )
 
     return queryResult[0]
 }
@@ -92,7 +110,12 @@ export async function resolve_blockchain_transactions(
 
     const { direction, limit } = processPaginationArgs(args)
 
-    const selectionSet = getNodeSelectionSetForConnection(info)
+    const useArchive = useTransactionsArchive(args.archive, context)
+    const { selectionSet, requireBocParsing } = upgradeSelectionForBocParsing(
+        useArchive,
+        getNodeSelectionSetForConnection(info),
+        transactionArchiveFields,
+    )
     const returnExpression = config.transactions.buildReturnExpression(
         selectionSet,
         context,
@@ -108,22 +131,26 @@ export async function resolve_blockchain_transactions(
         LIMIT ${limit}
         RETURN ${returnExpression}
     `
-    const queryResult = (await context.services.data.query(
-        required(context.services.data.transactions.provider),
-        {
-            text: query,
-            vars: params.values,
-            orderBy: [
-                {
-                    path: "chain_order",
-                    direction: "ASC",
-                },
-            ],
-            request: context,
-            traceSpan,
-            archive: args.archive ?? undefined,
-        },
-    )) as BlockchainTransaction[]
+    const queryResult = await parseTransactionBocsIfRequired(
+        requireBocParsing,
+        context,
+        (await context.services.data.query(
+            required(context.services.data.transactions.provider),
+            {
+                text: query,
+                vars: params.values,
+                orderBy: [
+                    {
+                        path: "chain_order",
+                        direction: "ASC",
+                    },
+                ],
+                request: context,
+                traceSpan,
+                archive: useArchive,
+            },
+        )) as BlockchainTransaction[],
+    )
 
     return (await processPaginatedQueryResult(
         queryResult,
@@ -173,7 +200,12 @@ export async function resolve_account_transactions(
 
     const { direction, limit } = processPaginationArgs(args)
 
-    const selectionSet = getNodeSelectionSetForConnection(info)
+    const useArchive = useTransactionsArchive(args.archive, context)
+    const { selectionSet, requireBocParsing } = upgradeSelectionForBocParsing(
+        useArchive,
+        getNodeSelectionSetForConnection(info),
+        transactionArchiveFields,
+    )
     const returnExpression = config.transactions.buildReturnExpression(
         selectionSet,
         context,
@@ -189,22 +221,26 @@ export async function resolve_account_transactions(
         LIMIT ${limit}
         RETURN ${returnExpression}
     `
-    const queryResult = (await context.services.data.query(
-        required(context.services.data.transactions.provider),
-        {
-            text: query,
-            vars: params.values,
-            orderBy: [
-                {
-                    path: "chain_order",
-                    direction: "ASC",
-                },
-            ],
-            request: context,
-            traceSpan,
-            archive: args.archive ?? undefined,
-        },
-    )) as BlockchainTransaction[]
+    const queryResult = await parseTransactionBocsIfRequired(
+        requireBocParsing,
+        context,
+        (await context.services.data.query(
+            required(context.services.data.transactions.provider),
+            {
+                text: query,
+                vars: params.values,
+                orderBy: [
+                    {
+                        path: "chain_order",
+                        direction: "ASC",
+                    },
+                ],
+                request: context,
+                traceSpan,
+                archive: useArchive,
+            },
+        )) as BlockchainTransaction[],
+    )
 
     return (await processPaginatedQueryResult(
         queryResult,