+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.60.0] - 2023-06-07
+
+### New
+
+- Added BOC parsing for transactions, messages and blocks in `archive` database when user selects
+  non parsed fields.
+
 ## [0.59.0] - 2023-06-07
 
 ### New